# Conflicts:
#	main.cpp

+++ net_work_stream.cpp
@@ -159,14 +159,14 @@ void WSWorkStream::PopData(FXNET::CNetStreamPackage& refPackage)
 		unsigned int dwHeaderLength = 0;
 		GetWSMsgLength((char*)this->GetData(), this->GetSize(), qwMsgLength, dwHeaderLength);
 		refPackage.~CNetStreamPackage();
-		new (&refPackage) FXNET::CNetStreamPackage((char*)m_btData + dwHeaderLength, qwMsgLength);
+		new (&refPackage) FXNET::CNetStreamPackage((char*)m_btData + dwHeaderLength, static_cast<unsigned int>(qwMsgLength));
 	
-		this->INetWorkStream::PopData(dwHeaderLength + qwMsgLength);
+		this->INetWorkStream::PopData(static_cast<unsigned int>(dwHeaderLength + qwMsgLength));
 		return;
 	}
 
 	refPackage.~CNetStreamPackage();
-	new (&refPackage) FXNET::CNetStreamPackage((char*)m_btData, this->GetSize());
+	new (&refPackage) FXNET::CNetStreamPackage((char*)m_btData, static_cast<unsigned int>(this->GetSize()));
 
 	this->INetWorkStream::PopData(this->GetSize());
 }
